pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lexile-corpus-tuner"
version = "0.1.0"
description = "Lexile-inspired corpus tuner with CLI for analyzing and rewriting documents."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Lexile Corpus Tuner Maintainers"}
]
dependencies = [
    "typer>=0.9.0",
    "PyYAML>=6.0",
    "numpy>=1.23",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.0",
]
lexile-v2 = [
    "tensorflow>=2.10.0",
    "joblib>=1.1.0",
]

[project.scripts]
lexile-tuner = "lexile_corpus_tuner.cli:main"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
testpaths = ["tests"]
